requirements.txt
streamlit>=1.22.0
pillow>=9.5.0
google-cloud-aiplatform>=1.25.0
requests>=2.28.0